refactor(mock-attestation): address review
- Move stamp_expiry_on_legacy_mocks (+ its test) into the v3_13_0_state
  migration module, since it is migration-only (netrome).
- Simplify submitted_attestation_landed docstring: a changed stored expiry
  is enough to conclude the submit landed; drop the confusing 'identity'
  wording (the legacy equality fallback is covered by TODO(#3786)).
